Correctness fixes: fractional workouts, distance fallback, route-only detect, overlapping-route dedup, per-window step source
Findings #7-#11 from the code review.

#7 A fractional value no longer 422s the whole ~15 MB sync: Workout
total_distance_meters/energy/steps accept float on the wire (Health Connect
sends fractional distances) and are rounded to int at ingest (_round_or_none),
so the DB columns and response models stay int.

#8 _activity_entries falls back to the matched workout's distance/duration when
the detected session's is None/0 (its distance source was dropped by dedup), so
a recorded run no longer contributes 0 km to weekly/summary stats.

#9 POST /detect covers route-only athletes (DIY routes, no wearable HR): a
concrete athlete_id runs unconditionally, and the all-athletes sweep unions
HeartRateSample + RouteTrack athlete ids instead of HR-only.

#10 run_detection_for_athlete no longer turns two overlapping DIY routes into
two sessions (double-counting distance): it tracks emitted route windows and
skips a route overlapping one (routes ordered by start_time for determinism).

#11 Detection picks the primary step/distance source PER SESSION WINDOW instead
of once globally, so a window the global source didn't cover (tracker charging)
no longer zeroes cadence and drops a real session. DETECTION_VERSION -> v3;
removed the now-unused _minute_range helper.

Verified: per-window source retains the session a global pick would drop (and
still doesn't double-count across sources); fractional workout ingests 201 and
stores rounded ints; matched-session distance falls back to the workout;
route-only /detect returns a session; overlapping routes collapse to one.

+++ schemas.py
@@ -38,9 +38,12 @@ class Workout(BaseModel):
     start_time: datetime
     end_time: datetime
     duration_seconds: int
-    total_distance_meters: int | None = None
-    total_energy_kcal: int | None = None
-    total_steps: int | None = None
+    # Accept floats on the wire: Health Connect distances/energies are often
+    # fractional, and Pydantic would otherwise 422 the WHOLE sync over one
+    # fractional value. Stored rounded to int at ingest (see _round_or_none).
+    total_distance_meters: float | None = None
+    total_energy_kcal: float | None = None
+    total_steps: float | None = None
 
 
 class TrackPoint(BaseModel):
